Default generation of CMake config packages to ON
Problem
-------

When consuming this project as a package, consumers will receive
errors when executing `find_package(beman.exemplar)`. This is
because by default the required CMake files are not installed with
the package.

Solution
--------

Bump the version of the infra subdirectory to use a new release
of the `beman-install-library` CMake module that provides
support for `BEMAN_EXEMPLAR_INSTALL_CONFIG_FILE_PACKAGE`
and `BEMAN_INSTALL_CONFIG_FILE_PACKAGES` options.

This change is not anticipated to affect `FetchContent` or other
kinds of vendored workflows since installation of libraries is
primarily a packaging activity. Even in that case, it is
harmless to consumers to install CMake modules that are not used.

Also included in this change is improvements to the
`beman.exemplar` README to note the existence of the
`BEMAN_EXEMPLAR_INSTALL_CONFIG_FILE_PACKAGE` option.

Documentation in `README.md` was also updated to reflect the
new installation behavior for `cmake --workflow --preset [...]`
workflows.